Get and save directories as JSON data
